MouseInput and GameState.Update
Added functions to hide and grab the cursor when desired, for free-look 1st-person style mouse control. Also updated the delta value provided to all update methods and modified TestState's update method to use the delta for smooth motion.

+++ src/com/degauvendesign/interstella/GameState.java
@@ -8,7 +8,7 @@ public interface GameState {
 	
 	public void input(Window window, MouseInput mouseInput);
 
-	public void update(Game game, double delta, MouseInput mouseInput);
+	public void update(Game game, float delta, MouseInput mouseInput);
 	
 	public void dispose(Game game);
 }
